Added Documentation And Removed Debug Logs

+++ Assets/General Player/PlayerHealth.cs
@@ -97,8 +97,7 @@ public class PlayerHealth : MonoBehaviour
     private void UpdateHealthLabel(bool damaged = false)
     {
         hudRef.SetHealthLabel(mainHealth.ToString());
-        //Debug.Log(mainHealth + "    " + maxMainHealth + "    " + mainHealth / maxMainHealth);
-        hudRef.SetHealthBarFill((float)mainHealth / (float)maxMainHealth);
+        hudRef.SetHealthBarFill(mainHealth / (float)maxMainHealth);
         if (damaged) hudRef.EnableDamageGradient();
     }
 }
